[SNAPPYDATA] Allow configurable MemoryManager
To plugin SnappyData's Unified Memory Manager (SnappyUnifiedMemoryManager), the
MemoryManager is made configurable using "spark.memory.manager" configuration property
that is set to the SnappyData's manager in embedded mode.

[SNAP-1256] (#41)
set the memory manager as spark's UnifiedMemoryManager, if spark.memory.manager is set as default

[SNAP-2084] (#86)
If SnappyUMM is found in classpath , SparkEnv will assign the memory manager to SnappyUMM.
If user has explicitly set the memory manager that will take precedence.