[SNAP-3165] Instantiating snappy session only when catalogImplementation (#191)
is in-memory which running pyspark shell.

## What changes were proposed in this pull request?

We are initializing `SparkSession` as well as `SnappySession` while starting pyspark shell.
`SparkSession` and `SparkContext`were always initialized with hive support enable
 irrespective of value of `spark.sql.catalogImplementation` config. 

With these changes, we are checking the value of `spark.sql.catalogImplementation` and
hive support is not enabled when the value of above-mentioned property is set to
 `in-memory` explicitly. 

SnappySession will be only initialized when catalog implementation is set to `in-memory`
to avoid failure reported in SNAP-3165. 

Later we can provide support for hive catalog implementation for python with SnappySession.