[SNAP-3270] removing streaming query listener in finalize block (#195)
## What changes were proposed in this pull request?

For streaming UI, `SnappyStreamingQueryListener` is registered on listener bus
at the time of creating `SnappySession`. However, this listener is never removed
from the listener bus. Hence even if the `SnappySession` is collected by GC,
`SnappyStreamingQueryListener` is left orphan on the listener bus which is never
eligible for GC collection.

To fix this we are removing the listener from the listener bus when the session
the instance is getting collected by GC (i.e. in finalize method) which will make
the listener instance eligible for GC during the next GC cycle.

It should be OK if the listener instance gets collected in the next GC cycle as the
the memory footprint of the listener object is not big.

Another possible place to remove the listener in `close` method of the session,
however close method of session is not required to be closed explicitly.

## How was this patch tested?

Reproduced the issue by running the following code as part of a snappy job: 

```
while(true){
      session.newSession()
}
```

Collected histogram of leader process using `jmap` and observed that instances of
`SnappyStreamingQueryListener` is increasing indefinitely and never garbage
collected whereas `SnappySession` instances are garbage collected: 

`jmap -histo:live <leader pid>|grep "SnappySession\|SnappyStreamingQueryListener"`

Followed the same steps after applying the changes and noticed that
 `SnappyStreamingQueryListener` instances are garbage collected.

--- 

- `SnappyStreamingQueryListenerSuite` is passing now. Added a call to `finalize`
method in `StreamingQueryListenerSuite` to get this change tested.