Fix for [SNAP-3269]
The byte array in the BufferHolder used to create an Unsafe JoinedRow is being reused,
while the UnsafeRow generated itself will get cached in the SnappyHashAggregate
(if the query plan is such that output of Window Expression is being used by Aggregate Function).
This causes the previous UnsafeRow objects cached to point to the current state of bytes
in the BufferHolder causing data corruption.

The fix is to copy the row generated from the WindowExpression.